+++ DutWrapper.TestRunner/DutWrapper.TestRunner.csproj
@@ -5,9 +5,9 @@
 
     <IsPackable>false</IsPackable>
 
-    <AssemblyVersion>1.7.0</AssemblyVersion>
+    <AssemblyVersion>1.7.1</AssemblyVersion>
 
-    <FileVersion>1.7.0</FileVersion>
+    <FileVersion>1.7.1</FileVersion>
 
     <NeutralLanguage>en-US</NeutralLanguage>
   </PropertyGroup>
